+++ ru.taximaxim.codekeeper.core/src/main/java/ru/taximaxim/codekeeper/core/schema/AbstractSchema.java
@@ -259,7 +259,7 @@ public abstract class AbstractSchema extends PgStatement implements ISchema {
      *
      * @return {@link #collations}
      */
-    public Collection<PgCollation> getColletions() {
+    public Collection<PgCollation> getCollations() {
         return Collections.unmodifiableCollection(collations.values());
     }