+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 
 - Fixed an error when reading locale for COLLATION objects.
+- Fixed a bug in generating the migration script for COLLATION.
 
 ## [8.7.0] - 2023-09-12